requirements.txt
# Core ML and Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0

# Text Processing and NLP
nltk>=3.7
spacy>=3.4.0
textblob>=0.17.0
wordcloud>=1.8.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Model Training and Evaluation
tensorboard>=2.9.0
wandb>=0.12.0
optuna>=3.0.0

# Data Processing
tqdm>=4.64.0
joblib>=1.1.0
h5py>=3.7.0

# Web Framework (for demo/API)
flask>=2.1.0
flask-cors>=3.0.10

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=0.19.0
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code Quality
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Jupyter and Development
jupyter>=1.0.0
ipykernel>=6.0.0
